Update README.md
This repository hosts the code and resources for a machine learning project focused on sentiment analysis of diacritized Arabic text using Recurrent Neural Networks (RNNs), particularly leveraging the MARBERT model. The project aims to explore and enhance the capabilities of RNNs in handling the complex morphology of the Arabic language, which is enriched with diacritical marks. The codebase includes preprocessing scripts, model training and evaluation modules, and utilities for handling diacritized Arabic text.